// src/app/shared/models/slick-carousel/slick-carousel.config.ts
export interface SlickCarouselConfig {
    infinite?: boolean;
    arrows?: boolean;
    dots?: boolean;
    draggable?: boolean;
    autoplay?: boolean;
    inifinite?: boolean;
    slidesToShow?: number;
    slidesToScroll?: number;
    appendArrows?: any;
    prevArrow?: string;
    nextArrow?: string;
    repsonsive?: any;
}